pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-merge"
version = "0.0.1"
description = "Fuzzy Merging Algorithm using LLM"
authors = [
{ name = "Taha Ibrahim Siddiqui", email = "taha.i.siddiQ@gmail.com" },
{ name = "Asad Tariq", email = "asad@example.com" } # Add actual email
]
readme = "README.md"
license =  "MIT" 
requires-python = ">=3.9"
dependencies = [
"pandas>=1.5.0",
"numpy>=1.23.0",
"tqdm>=4.66.0",
"google-generativeai>=0.7.0"
]

classifiers = [
"Development Status :: 3 - Alpha",
"Intended Audience :: Developers",
"Intended Audience :: Science/Research",
"Programming Language :: Python :: 3",
"Programming Language :: Python :: 3.9",
"Programming Language :: Python :: 3.10",
"Programming Language :: Python :: 3.11",
"Programming Language :: Python :: 3.12",
"Operating System :: OS Independent",
"Topic :: Scientific/Engineering",
"Topic :: Software Development :: Libraries :: Python Modules"
]

[project.urls]
Homepage = "https://github.com/yourusername/ai-merge" # Update with real URL
Issues = "https://github.com/yourusername/ai-merge/issues" # Update with real URL
Documentation = "https://github.com/yourusername/ai-merge#readme"

[project.optional-dependencies]
dev = [
"pytest>=7.0.0",
"pytest-cov>=4.0.0",
"black>=23.0.0",
"ruff>=0.1.0",
"mypy>=1.0.0",
"pandas-stubs>=2.0.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]